Feature: New reducer role control
