+++ client/src/components/Layout.jsx
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { Outlet } from 'react-router-dom'
-import Navbar from './Navbar'
+import Navbar from './Navbar/Navbar'
 
 const Layout = () => {
   return (
