+++ client/src/pages/Info/Contacto.jsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import * as yup from 'yup'
-import useIntersectionObserver from '../components/hooks/useObserver'
+import useIntersectionObserver from '../../components/hooks/useObserver'
 const Contacto = () => {
   const AnimatedComponent = useIntersectionObserver({
     threshold: 0.5,
