+++ client/src/pages/Into.jsx
@@ -1,0 +1,3 @@
+export const Into = () => {
+  return <>name</>
+}
