+++ client/src/router.jsx
@@ -1,14 +1,14 @@
 import { createBrowserRouter } from 'react-router-dom'
 import LandingPage from '@/pages/Landing'
-import Ofertas from './pages/Ofertas'
-import Empresas from './pages/Empresas'
-import Noticias from './pages/Noticias'
-import Acerca from './pages/Acerca'
-import Contacto from './pages/Contacto'
-import Login from './pages/Login'
-import Register from './pages/Register'
+import Ofertas from './pages/Info/Ofertas'
+import Empresas from './pages/Community/Empresas'
+import Noticias from './pages/Info/Noticias'
+import Acerca from './pages/Info/Acerca'
+import Contacto from './pages/Info/Contacto'
+import Login from './pages/Auth/Login'
+import Register from './pages/Auth/Register'
 import Layout from './components/Layout'
-import ErrorPage from './pages/ErrorPage'
+import ErrorPage from './pages/404'
 export const router = createBrowserRouter([
   {
     path: '/',
